requirements.txt
langchain>=0.0.335
openai>=1.2.3
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.1.1
spacy>=3.7.2
sqlalchemy>=2.0.22
apscheduler>=3.10.4
selenium>=4.14.0
python-dotenv>=1.0.0
pyautogen>=0.2.0
crewai>=0.1.20
azure-functions>=1.17.0
boto3>=1.28.44
numpy>=1.26.0
scikit-learn>=1.3.2
pytest>=7.4.3